scripts/openclaw_projects_index/requirements.txt
PyYAML==6.0.3
psycopg[binary]==3.2.10
sqlparse==0.5.3
